% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_utils.r
\name{addfilename}
\alias{addfilename}
\title{addfilename adds a filename to the autoresult csv file}
\usage{
addfilename(filename, tabfile, category, type, caption)
}
\arguments{
\item{filename}{the full path and filename for the file being added}

\item{tabfile}{the file to be added to, which is defined by
setuphtml found in aMSE_utils}

\item{category}{what HTML tab should it be displayed on?}

\item{type}{the type of addition, either a plot or table}

\item{caption}{the caption for the figure or table}
}
\value{
nothing but it does add a line to filen
}
\description{
addfilename is used to facilitate the produciton of
    the HTML results summary for a particular run. This depends
    upon a csv file containing the names of each file to be
    plotted or tabulated. This function adds a filename and the
    supporting caption and category, without one needing to
    remember the syntax.
}
\examples{
indir <- tempdir()
plotdir <- filenametopath(indir,"plots")
dirExists(plotdir,verbose=FALSE)
plottabfile <- setuphtml(plotdir,"example_only")
filename <- filenametopath(plotdir,"example.png")
png(filename=filename,width=7,height=4,units="in",res=300)
plot(runif(100),runif(100),type="p")
graphics.off()  # could use dev.off()
addfilename(filename=filename,tabfile=plottabfile,"A_category",
            type="plot",caption="Example Figure")
dir(plotdir)
}
