% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{addlist}
\alias{addlist}
\title{addlist write a list's bits to a file and adds that to resultTable.csv}
\usage{
addlist(inlist, filen, rundir, category = "any", caption = "")
}
\arguments{
\item{inlist}{A list object containing whatever a list will contain}

\item{filen}{either just the filename or the full path and filename}

\item{rundir}{full path to the directory to contain the results}

\item{category}{what HTML tab should it be added to? default="any"
obviously? you would want to change this.}

\item{caption}{the caption for the list object default = "",
This should not contain commas as this confuses the csv file.
But if you accidentally put some in they will be removed.}
}
\value{
nothing but it does add a line to resfile
}
\description{
addlist it is common to want to include the contents of a list 
    to the output tabs. addlist can include numeric constants, and vectors,
    2D matrices and data.frames but not arrays or lists within lists.
}
\examples{
library(makehtml)
indir <- tempdir()
rundir <- filenametopath(indir,"result")
dirExists(rundir,verbose=FALSE)
resfile <- setuphtml(rundir=rundir)
filename <- filenametopath(rundir,"example.png") # must be a png file
png(filename=filename,width=7,height=4,units="in",res=300)
plot(runif(100),runif(100),type="p")
addplot(filen=filename,rundir=rundir,category="EG_Figure",caption="Eg Figure") 
filename <- "example2.png" # note png requires full path, addplot doesn't
png(filename=filenametopath(rundir,filename),width=7,height=4,units="in",
    res=300)
plot(runif(100),runif(100),type="l",lwd=2,col="red")
addplot(filen=filename,rundir=rundir,category="EG_Figure2",
        caption="Eg Figure2") 
x <- rnorm(10,mean=5,sd=1); y <- as.data.frame(cbind(x,x))
z <- list(a=5,b="hello",c=rnorm(10,mean=5,sd=1))
eg <- list(x=x,y=y,z=z)
addlist(eg,filen="example.txt",rundir=rundir,category="listadded",
        caption="List")
runnotes= c("can normally have both plots on one tab",
            "but using tempdir behaves flakily")
make_html(rundir=rundir,packagename="makehtml",runnotes=runnotes)
}
