% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{addplot}
\alias{addplot}
\title{addplot adds a plot's filename to the autoresult csv file}
\usage{
addplot(filen, resdir, category = "any", caption = "")
}
\arguments{
\item{filen}{full path and filename for file being added}

\item{resdir}{full path to the directory to contain the results}

\item{category}{what HTML tab should it be added to? default="any"
obviously? you would want to change this.}

\item{caption}{the caption for the figure or table, default = "",
This should not contain commas as this confuses the csv file.
But if you accidentally put some in they will be removed.}
}
\value{
nothing but it does add a line to resfile
}
\description{
addplot is used to facilitate the production of
    the HTML results summary for a particular run. If the filen remains 
    empty, then addplot does nothing. However, if filen is longer than "",
    then addplot adds the plot's filename and the supporting caption and 
    category to resultTable.csv that lists all files to be included in 
    the local website. If no category is added explicitly then the local 
    webpage will have an 'any' tab containing these unloved results.
}
\examples{
indir <- tempdir()
resdir <- filenametopath(indir,"result")
dirExists(resdir,verbose=FALSE)
resfile <- setuphtml(resdir=resdir)
filename <- "example.png" # must be a png file
png(filename=filename,width=7,height=4,units="in",res=300)
plot(runif(100),runif(100),type="p")
addplot(filen=filename,resdir=resdir,"A_category",caption="Example Figure")  
dir(resdir)
}
