% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{addtext}
\alias{addtext}
\title{addtext literally enables the addition of a text block to a tab}
\usage{
addtext(txt, rundir, filename, category = "text")
}
\arguments{
\item{txt}{the vector of character strings to be added as a text block.}

\item{rundir}{full path to the directory to contain the results}

\item{filename}{the name of text or CSV file to store ready for inclusion in the 
htmlfile defined by category. Be sure to include the filetype .txt}

\item{category}{what HTML tab should it be added to? default="text"}
}
\value{
nothing but it does add text to a tab within an html file in rundir. 
    It does not make a note in resultTable.csv
}
\description{
addtext provides a solution for when one wants to add a block
    of explanatory text to a tab in the internal website created by make_html.
}
\examples{
indir <- tempdir()
rundir <- filenametopath(indir,"result")
dirExists(rundir,verbose=FALSE)
resfile <- setuphtml(rundir)
filen <- "example.txt"
# make a character vector
egtxt <- c("This is example text. Of course one could put anything in",
           "here as commentary to other results. This function is ",
           "here for completeness, though I use it very little.")
addtext(egtxt,filen=filen,rundir=rundir,category="ExampleText")
dir(rundir) # examine the resfile and the example.csv files.
make_html(rundir=rundir,packagename="makehtml") 
}
