% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{cleanrundir}
\alias{cleanrundir}
\title{cleanrundir removes all html, png, and css files from rundir}
\usage{
cleanrundir(rundir, verbose = TRUE, abmse = TRUE)
}
\arguments{
\item{rundir}{a specific directory used for an analytical run}

\item{verbose}{default=TRUE. This assumes no action unless an explicit 
Y or y is entered at the prompt. If set to TRUE it will delete the
files with no explicit prompt.}

\item{abmse}{should a specific set of files be deleted (see description)}
}
\value{
invisibly the remaining contents of rundir as a vector
}
\description{
cleanrundir removes all html, png, and css files from 
    rundir. In addition, if abmse=TRUE, the default, then up to eight
    specific files are removed. This function is to replace the use of
    cleanslate in setuphtml, which was more clumsy. The special aMSE 
    files to be removed are: "final_harvestR.csv","glb.RData","hsargs.txt",
    "HSstats.RData","popdefs.csv","propertyDD.csv","resultTable.csv",
    "zonebiology.csv".
}
\examples{
\dontrun{
  print("wait on an example file")
}
}
