% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{dirExists}
\alias{dirExists}
\title{dirExists: Checks for the existence of a directory}
\usage{
dirExists(indir, make = TRUE, verbose = TRUE)
}
\arguments{
\item{indir}{a character string containing the name of the directory
whose existence is to be checked before it is created if it
does not already exist.}

\item{make}{if the directory does NOT exist should it be created.
default = TRUE; if make=FALSE and a directory does not exist
a warning will be given to the console.}

\item{verbose}{default=TRUE, prints directory status to the console,
If make is set to FALSE and a directory does not exist a
warning will always be given.}
}
\value{
a message to the screen if the directory exists or is
    created; if make is TRUE then it also creates the directory as
    listed in 'indir'.
}
\description{
dirExists: does a directory exist? It uses dir.exists
    and reports existence if already present and uses dir.create
    it it does not exist, but avoids the warning message is one
    already exists. The option of not creating a new directory is
    also present. This uses 'recursive=TRUE' inside the dir.create, so
    one should be able to create directories as deeply down a path as
    wished.
}
\examples{
indirect <- getwd()
dirExists(indirect)
}
