% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{endmakehtml}
\alias{endmakehtml}
\title{endmakehtml write the syntax of the website generation to console}
\usage{
endmakehtml()
}
\value{
nothing but it does write some code syntax to the console
}
\description{
endmakehtml writes to the console the syntax of the final set 
    of commands needed to generate the internal website of results. The idea 
    is that one would then copy those lines to the end of one's own code to
    complete the output of results. Its functions is simply to facilitate 
    remembering the syntax.
}
\examples{
endmakehtml()
}
