% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_utils.r
\name{filenametopath}
\alias{filenametopath}
\title{filenametopath safely add a filename to a path}
\usage{
filenametopath(inpath, infile)
}
\arguments{
\item{inpath}{the path to be analysed}

\item{infile}{the filename to be added to the path 'inpath'}
}
\value{
the completed filename or extended path
}
\description{
filenametopath add a filename to a path safely, using
    pathtype to get the seperator and then checks the end character.
    If the separator is nothing or a '/' or a '//' then it adds to
    the path appropriately. Without this one can unwittingly include
    extra separators or none at all.
}
\examples{
indir <- tempdir()
infile <- "control.csv"
filenametopath(indir,infile)
}
