% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{getextension}
\alias{getextension}
\title{getextension gets the 3 letter file extension from a filename}
\usage{
getextension(filename)
}
\arguments{
\item{filename}{the filename, with path attached, whose extension
needs to be determined.}
}
\value{
a character string of either 'table' or 'plot'
}
\description{
getextension is a utility function that is used with
    logfilename to determine whether one is dealing with a csv or a
    png file. depending on which it will return either a 'table' or
    'plot' value. Any other extension type will throw an error and
    stop execution.
}
\examples{
filen <- "not_a_real_file.png"
getextension(filen)
filen <- "another_unreal_file.csv"
getextension(filen)
}
