% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{htmltable}
\alias{htmltable}
\title{htmltable generates the html to print out a table}
\usage{
htmltable(inmat, filename, caption)
}
\arguments{
\item{inmat}{the 2D matrix or data.frame to be printed}

\item{filename}{the filename to which to add the html, defined by
makehtml}

\item{caption}{the caption text placed at the top of the table}
}
\value{
nothing but it does add some html to the input filename
}
\description{
htmltable generates the required html code to add a
    table to the website of the results. This requires the
    addfilename function to include both a category biology,
    productivity, etc, and a type, which is currently limited to
    plot or table.
}
