% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{logfilename}
\alias{logfilename}
\title{logfilename adds a filename to the autoresult csv file}
\usage{
logfilename(filename, resfile, category = "any", caption = "", type = "")
}
\arguments{
\item{filename}{the full path and filename for the file being added}

\item{resfile}{the file to be added to, which is defined by
setuphtml found in aMSE_utils}

\item{category}{what HTML tab should it be added to? default="any"}

\item{caption}{the caption for the figure or table, default = "",
This should not contain commas as this confuses the csv file.
But if you accidentally put some in they will be removed.}

\item{type}{allows one to override the plot and table options.
Currently the alternative is bigtable}
}
\value{
nothing but it does add a line to resfile
}
\description{
logfilename is used to facilitate the produciton of
    the HTML results summary for a particular run. This depends
    upon a csv file containing the names of each file to be
    plotted or tabulated. This function adds a filename and the
    supporting caption and category, without one needing to
    remember the syntax. If no category is added explicitly then
    the local webpage will have an 'any' tab containing these
    unloved results. logfilename would not usually be called 
    independently of addtable or addplot but it is exported for 
    completeness.
}
\examples{
indir <- tempdir()
rundir <- filenametopath(indir,"result")
dirExists(rundir,verbose=FALSE)
resfile <- setuphtml(rundir)
filename <- filenametopath(rundir,"example.png")
png(filename=filename,width=7,height=4,units="in",res=300)
plot(runif(100),runif(100),type="p")
dev.off()  
logfilename(filename=filename,resfile=resfile,"A_category",
            caption="Example Figure")
dir(rundir)
}
