% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{make_html}
\alias{make_html}
\title{make_html create HTML files to view results in a browser.}
\usage{
make_html(
  replist = NULL,
  rundir = NULL,
  datadir = NULL,
  controlfile = NULL,
  datafile = NULL,
  hsfile = NULL,
  width = 500,
  openfile = TRUE,
  runnotes = NULL,
  verbose = TRUE,
  packagename = "mainpackage",
  htmlname = "htmlname"
)
}
\arguments{
\item{replist}{Object created by a run, can be NULL}

\item{rundir}{Directory where a particular run's files, including 
any results, as tables and plots, and any other files, are all 
held. Cannot be NULL.}

\item{datadir}{full path to the data directory, if one is used}

\item{controlfile}{the character name of the control file used.}

\item{datafile}{the character name of the saudata data file.}

\item{hsfile}{the character name of the harvest strategy file, default=NULL}

\item{width}{Width of plots (in pixels). Default = 500}

\item{openfile}{Automatically open index.html in default browser?}

\item{runnotes}{Add additional notes to home page.}

\item{verbose}{Display more info while running this function?}

\item{packagename}{name of the main package being used in the 
analysis. default='aMSE'. This is described on the home page}

\item{htmlname}{first name of the html files generated, default='aMSE'}
}
\description{
make_html writes a set of HTML files with tabbed
    navigation, with each tab containing the results relating to a
    given set of results or diagnostics from a particular aMSE run.
    This code was borrowed from Ian Taylor's r4ss, but has been
    extensively modified to improve both the css (see write_css)
    and the HTML. By default, this function will look in the
    results directory where PNG and CSV files were created for a
    resultTable.csv file. HTML files are written to link to these plots
    and put in the same directory. Output now includes the control, data,
    and HS files names. This was written to facilitate the presentation of
    result for an MSE, hence some of the names. But, of course, it can be
    used for any set of results.
}
\examples{
library(makehtml)
indir <- tempdir()
rundir <- filenametopath(indir,"result")
dirExists(rundir,verbose=FALSE)
resfile <- setuphtml(rundir=rundir)
start <- Sys.time()
addnotes <- "No warnings reported."
filename <- filenametopath(rundir,"example.png") # must be a png file
png(filename=filename,width=7,height=4,units="in",res=300)
plot(runif(100),runif(100),type="p")
addplot(filen=filename,rundir=rundir,"A_category",caption="Example Figure")
filen <- "example.csv"
egtable <- matrix(rnorm(25,0,1),nrow=5,ncol=5)
addtable(egtable,filen=filen,rundir=rundir,"ExampleTable",
            caption="An example Table")
# make a character vector
egtxt <- c("This is example text. Of course one could put anything in",
           "here as commentary to other results. This function is ",
           "here for completeness, though I use it very little.")
addtext(egtxt,filen=filen,rundir=rundir,category="ExampleText")
dir(rundir)
runnotes <- c(paste0("RunTime = ",Sys.time() - start),
              paste0("replicates = ",100),paste0("years projected = ",30),
              addnotes)
make_html(replist=NULL,rundir=rundir,datadir=rundir,hsfile="anyname",
          width=500,openfile=TRUE,runnotes=runnotes,
          packagename="makehtml",htmlname="example")
}
\author{
Originally Ian Taylor, modified by Malcolm Haddon
}
