% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{plotfilename}
\alias{plotfilename}
\title{plotfilename simplifies the production of a}
\usage{
plotfilename(plotname, runlabel, resdir)
}
\arguments{
\item{plotname}{name of plot as character string}

\item{runlabel}{name of particular run as character string}

\item{resdir}{the results directory as character string}
}
\value{
a filename to be included into resultTable.csv
}
\description{
plotfilename simplifies the production of a
}
\examples{
resdir <- tempdir()
plotfilename("Schaefer_FisheryPlot","runone",resdir)
}
