% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{setuphtml}
\alias{setuphtml}
\title{setuphtml initiates csv files listing results to be included}
\usage{
setuphtml(rundir, cleanslate = FALSE)
}
\arguments{
\item{rundir}{full path to the directory to contain the results}

\item{cleanslate}{this is now deprecated and has been replaced by the 
function cleanrundir, which is safer and clearer.}
}
\value{
invisibly, the full path to the resfile, after creating the file in 
    rundir and potentially deleting all previous html, png, and .css files 
    contained in rundir
}
\description{
setuphtml always initiates the resultTable.csv file used to 
    contain the filenames, captions, and categories of the plots and 
    tables to be included in the html results. The format of the csv 
    file is to have column names of file, caption, category, and
    timestamp. Then, each plot and table is included with an entry
    for each column.
}
\examples{
indir <- tempdir()
rundir <- filenametopath(indir,"results")
dirExists(rundir,verbose=FALSE)
resfile <- setuphtml(rundir)
dir(rundir)
}
