% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{write_css}
\alias{write_css}
\title{write_css generates a CSS file used by all html files}
\usage{
write_css(plotdir)
}
\arguments{
\item{plotdir}{the directory within the run directory that contains
all the plot results}
}
\value{
nothing but it does generate a .css file in the plotdir
}
\description{
write_css generates a cascading style sheet that will
    be used by each separate html files made for each category of
    results. The origin came from Ian Taylor although I have
    modified the styles and ensure correct Version 3 CSS
}
