% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{write_head}
\alias{write_head}
\title{write_head adds the <head> tag to  each html file in results}
\usage{
write_head(htmlfile)
}
\arguments{
\item{htmlfile}{the particular html file being worked on. This is
defined within the make_html function.}
}
\value{
nothing but it does add the <head> tag to each html file
}
\description{
write_head adds the head tag to each html file.
    It links to the css file rather than writing the complete
    css code to every file. There were missing elements which are
    now in place, so the code is now valid HTML5.
}
