% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{addtable}
\alias{addtable}
\title{addtable adds a table to the output}
\usage{
addtable(intable, filen, resfile, category = "any", caption = "")
}
\arguments{
\item{intable}{the table or data.frame to be saved and output}

\item{filen}{the full path and filename being given to the table}

\item{resfile}{the file to be added to, which is defined by
setuphtml found in aMSE_utils}

\item{category}{what HTML tab should it be added to? default="any"}

\item{caption}{the caption for the figure or table, default = ""}
}
\value{
nothing but it does add a line to resfile and saves a csv 
    file to resdir
}
\description{
addtable saves a table as a csv file into the resdir.
    Then it logs the filename in the the csv file containing the 
    names of each file to be plotted or tabulated. This function 
    saves having to combined the filename with the resdir, and then
    does the filename logging for you. If no category is added 
    explicitly then the local webpage will have an 'any' tab 
    containing these unloved results.
}
\examples{
indir <- tempdir()
resdir <- filenametopath(indir,"result")
dirExists(resdir,verbose=FALSE)
resfile <- setuphtml(resdir,"example_only")
filen <- "example.csv")
egtable <- matrix(rnorm(25,0,1),nrow=5,ncol=5)
addtable(egtable,filen=filen,resfile=resfile,"A_category",
            caption="An example Table")
dir(resdir) # examine the resfile and the example.csv files.
}
