% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{make_html}
\alias{make_html}
\title{make_html create HTML files to view results in a browser.}
\usage{
make_html(
  replist = NULL,
  rundir = NULL,
  width = 500,
  openfile = TRUE,
  runnotes = NULL,
  verbose = TRUE
)
}
\arguments{
\item{replist}{Object created by an aMSE run}

\item{rundir}{Directory where a particular run's files,
including the region files, the results, as tables and plots,
and any other files, are all held. It will always contain at
least the 'data' and 'results' sub-directories.}

\item{width}{Width of plots (in pixels). Default = 500}

\item{openfile}{Automatically open index.html in default browser?}

\item{runnotes}{Add additional notes to home page.}

\item{verbose}{Display more info while running this function?}
}
\description{
make_html writes a set of HTML files with tabbed
    navigation, with each tab containing the results relating to a
    given set of results or diagnostics from a particular aMSE run.
    This code was borrowed from Ian Taylor's r4ss, but has been
    extensively modified to improve both the css (see write_css)
    and the HTML. By default, this function will look in the
    results directory where PNG files were created for a
    resultTable_runname.csv file with the name 'runname' written
    added by aMSE. HTML files are written to link to these plots
    and put in the same directory.
}
\author{
Originally Ian Taylor, modified by Malcolm Haddon
}
