% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{make_html}
\alias{make_html}
\title{make_html create HTML files to view results in a browser.}
\usage{
make_html(
  replist = NULL,
  resdir = NULL,
  width = 500,
  openfile = TRUE,
  runnotes = NULL,
  verbose = TRUE,
  packagename = "aMSE",
  htmlname = "aMSE"
)
}
\arguments{
\item{replist}{Object created by a run, can be NULL}

\item{resdir}{Directory where a particular run's files, including 
any results, as tables and plots, and any other files, are all 
held. Cannot be NULL.}

\item{width}{Width of plots (in pixels). Default = 500}

\item{openfile}{Automatically open index.html in default browser?}

\item{runnotes}{Add additional notes to home page.}

\item{verbose}{Display more info while running this function?}

\item{packagename}{name of the main package being used in the 
analysis. default='aMSE'. This is described on the home page}

\item{htmlname}{first name of the html files generated, default='aMSE'}
}
\description{
make_html writes a set of HTML files with tabbed
    navigation, with each tab containing the results relating to a
    given set of results or diagnostics from a particular aMSE run.
    This code was borrowed from Ian Taylor's r4ss, but has been
    extensively modified to improve both the css (see write_css)
    and the HTML. By default, this function will look in the
    results directory where PNG and CSV files were created for a
    resultTable.csv file. HTML files are written to link to these plots
    and put in the same directory.
}
\author{
Originally Ian Taylor, modified by Malcolm Haddon
}
