% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_utils.r
\name{setuphtml}
\alias{setuphtml}
\title{setuphtml initiates csv files lsiting results to be included}
\usage{
setuphtml(pldir, runname)
}
\arguments{
\item{pldir}{full path to the directory to contain the plots}

\item{runname}{the name of the particular run being summarized.}
}
\value{
full path to the plottabfile. creating the file in plotdir
}
\description{
setuphtml initiates the csv file used to contain the
    filenames, captions, and categories of the plots and tables to
    be included in the html results. The format of the csv file
    is to have column names of file, caption, category, and
    timestamp. Then, each plot and table is included with an entry
    for each column.
}
\examples{
\dontrun{
indir <- tempdir()
plotdir <- filenametopath(indir,"plots")
dirExists(plotdir,verbose=TRUE)
plottabfile <- setuphtml(plotdir,"example_only")
dir(plotdir)
}
}
