% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{setuphtml}
\alias{setuphtml}
\title{setuphtml initiates csv files listing results to be included}
\usage{
setuphtml(rundir)
}
\arguments{
\item{rundir}{full path to the directory to contain the results}
}
\value{
invisibly, the full path to the resfile, after creating the file in 
    rundir
}
\description{
setuphtml always initiates the resultTable.csv file used to 
    contain the filenames, captions, and categories of the plots and 
    tables to be included in the html results. The format of the csv 
    file is to have column names of file, caption, category, and
    timestamp. Then, each plot and table is included with an entry
    for each column.
}
\examples{
indir <- tempdir()
rundir <- filenametopath(indir,"results")
dirExists(rundir,verbose=TRUE)
resfile <- setuphtml(rundir)
dir(rundir)
}
