% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{setuphtml}
\alias{setuphtml}
\title{setuphtml initiates csv files lsiting results to be included}
\usage{
setuphtml(resdir, runname, cleanslate = TRUE)
}
\arguments{
\item{resdir}{full path to the directory to contain the plots}

\item{runname}{the name of the particular run being summarized.}

\item{cleanslate}{should the directory be emptied of results files first? 
Only files listed in resultTable_runname.csv are deleted. This means 
any data files put into the resdir will remain. default=TRUE}
}
\value{
full path to the resfile. creating the file in resdir
}
\description{
setuphtml initiates the csv file used to contain the
    filenames, captions, and categories of the plots and tables to
    be included in the html results. The format of the csv file
    is to have column names of file, caption, category, and
    timestamp. Then, each plot and table is included with an entry
    for each column.
}
\examples{
indir <- tempdir()
resdir <- filenametopath(indir,"results")
dirExists(resdir,verbose=FALSE)
resfile <- setuphtml(resdir,"example_only")
dir(resdir)
}
