% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makehtml_funs.R
\name{setuphtml}
\alias{setuphtml}
\title{setuphtml initiates csv files listing results to be included}
\usage{
setuphtml(resdir, cleanslate = FALSE)
}
\arguments{
\item{resdir}{full path to the directory to contain the results}

\item{cleanslate}{should the directory be emptied of all files first? 
All files in the directory will be deleted. default=FALSE. This is 
obviously a very powerful and potentially dangerous argument, hence 
it needs to be set =TRUE explicitly}
}
\value{
invisibly, the full path to the resfile, after creating the file in 
    resdir and potentially deleting all previous files contained in resdir
}
\description{
setuphtml always initiates the resultTable.csv file used to 
    contain the filenames, captions, and categories of the plots and 
    tables to be included in the html results. The format of the csv 
    file is to have column names of file, caption, category, and
    timestamp. Then, each plot and table is included with an entry
    for each column.
}
\examples{
indir <- tempdir()
resdir <- filenametopath(indir,"results")
dirExists(resdir,verbose=FALSE)
resfile <- setuphtml(resdir)
dir(resdir)
}
